Relocation register component, wire up with view, style tweaks
