+++ src/App.tsx
@@ -5,7 +5,7 @@ import { Projects } from './views/Projects';
 import { Home } from './views/Home';
 import { AppHeader } from "./components/AppHeader";
 import { Launch } from "./views/Launch";
-import { Register } from "./views/launch/register";
+import { Register } from "./views/dashboard/registerapp";
 import { Signin } from "./views/signin";
 import { Dashboard } from "./views/dashboard";
 
@@ -31,9 +31,9 @@ export const App: React.StatelessComponent = () => {
           <Route path="/" exact component={Home} />
           <Route path="/projects" component={Projects} />
           <Route exact path="/launch" component={Launch} />
-          <Route exact path="/launch/register" component={Register} />
           <Route exact path="/signin" component={Signin} />
           <Route exact path="/dashboard" component={Dashboard} />
+          <Route exact path="/dashboard/registerapp" component={Register} />
           <Route component={NotFound} />}
         </Switch>
       </Stack>
